+++ workalendar/tests/test_australia.py
@@ -59,6 +59,7 @@ class AustraliaCapitalTerritoryTest(AustraliaTest):
         holidays = self.cal.holidays_set(2013)
         self.assertIn(date(2013, 3, 11), holidays)
         self.assertIn(date(2013, 3, 30), holidays)  # Easter Saturday
+        self.assertIn(date(2013, 3, 31), holidays)  # Easter Sunday
         self.assertIn(date(2013, 4, 1), holidays)  # Easter Monday
         self.assertIn(date(2013, 6, 10), holidays)  # Queen's Bday
         self.assertIn(date(2013, 9, 30), holidays)  # Family & Community day
@@ -104,6 +105,25 @@ class AustraliaCapitalTerritoryTest(AustraliaTest):
         reconciliation_day = self.cal.get_reconciliation_day(2019)
         self.assertEqual(reconciliation_day, (date(2019, 5, 27),
                                               "Reconciliation Day"))
+
+    def test_full_2021(self):
+        holidays = self.cal.holidays_set(2021)
+        self.assertIn(date(2021, 1, 1), holidays)  # New Year's Day
+        self.assertIn(date(2021, 1, 26), holidays)  # Australia Day
+        self.assertIn(date(2021, 3, 8), holidays)  # Canberra Day
+        self.assertIn(date(2021, 4, 2), holidays)  # Good Friday
+        self.assertIn(date(2021, 4, 3), holidays)  # Easter Saturday
+        self.assertIn(date(2021, 4, 4), holidays)  # Easter Sunday
+        self.assertIn(date(2021, 4, 5), holidays)  # Easter Monday
+        # Anzac day displaced from 25th (SUN) to 26th
+        self.assertNotIn(date(2021, 4, 25), holidays)
+        self.assertIn(date(2021, 4, 26), holidays)
+
+        self.assertIn(date(2021, 5, 31), holidays)  # Reconciliation day
+        self.assertIn(date(2021, 6, 14), holidays)  # Queen's Birthday
+        self.assertIn(date(2021, 10, 4), holidays)  # Labour day
+        self.assertIn(date(2021, 12, 25), holidays)  # Xmas
+        self.assertIn(date(2021, 12, 26), holidays)  # Boxing Day
 
 
 class NewSouthWalesTest(AustraliaTest):